% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkList.R
\name{checkList}
\alias{checkList}
\title{Check List function}
\usage{
checkList(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
The List type.
}
\description{
The function checks that the input argument is a list containing either strings or GRranges.
}
\examples{
List_1 = list("A", "B", "C") # This is correct
List_2 = list(2, 1, 4) # This will raise an error
checkList(List_1)
checkList(List_2)
}
