% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTracks.R
\name{createTracks}
\alias{createTracks}
\title{Plot annotation track function}
\usage{
createTracks(
  dataList,
  track.names = NULL,
  region,
  normalise = FALSE,
  use.smooth = TRUE,
  win.smooth = 1001
)
}
\arguments{
\item{dataList}{A List object containing either files that can be read using the import function from rtracklayer or a list of GRanges, GAlignments, or GAlignmentPairs objects.}

\item{track.names}{A character vector of the same length as dataList containing the name to assign to each track.}

\item{region}{A GRanges object with a single entry defining the genomic region to visualise.}

\item{normalise}{A logical indicating whether to normalise the input data. Currently, only counts per million (CPM) normalisation is available.}

\item{use.smooth}{A logical indicating whether to apply a smoothing function (rolling mean of win.smooth window size).}

\item{win.smooth}{An positive integer defining the width of the window to use for the smoothing.}
}
\value{
A ggplot object.
}
\description{
The function extracts the annotated features from the TxDb object and distributes them on different plotting layers to prevent overlaps.
}
