% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChrSizes.R
\name{getChrSizes}
\alias{getChrSizes}
\title{Create SeqInfo object function}
\usage{
getChrSizes(genome)
}
\arguments{
\item{genome}{A character indicating a genome reference (e.g, "hg19", "mm10").}
}
\value{
A data.table object.
}
\description{
The function retrieves the chromosome size information for the genome in input and create the Seqinfo object.
}
\examples{
SeqInfo = getChrSizes("mm10")
}
