% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnno.R
\name{plotAnno}
\alias{plotAnno}
\title{Plot annotation track function}
\usage{
plotAnno(
  data = list(),
  reg.range = NULL,
  gene.height = 1,
  colorVar = "grey30",
  addID = TRUE,
  clean = FALSE,
  ...
)
}
\arguments{
\item{data}{A List object (the output of createAnnoTrack() ).}

\item{reg.range}{A vector of two integers, which define start and end for the plot left- and right-most borders.}

\item{gene.height}{A positive numeric value used as a multiplying factor for the gene height; the default is 1.}

\item{colorVar}{A character string indicating the colour of the gene lines/arrows; the default is grey.}

\item{addID}{A logical indicating whether to add a label to each gene. IMPORTANT: the data object must have a 'gene_name' or 'gene_id' column. If both are present, the former has priority.}

\item{clean}{A logical indicating whether to plot the frame around the annotation track.}

\item{...}{Other arguments passed on to ‘theme_bw()’.}
}
\value{
A ggplot object.
}
\description{
The function extracts the annotated features from the TxDb object and distributes them on different plotting layers to prevent overlaps.
}
